Moved installer code to admin_init
